Rename aggregation attribute to aggregation action.
