+++ graylog2-web-interface/src/views/components/aggregationwizard/actionConfiguration/MetricConfiguration.tsx
@@ -19,17 +19,18 @@ import { isEmpty } from 'lodash';
 import styled from 'styled-components';
 
 import AggregationWidgetConfig from 'views/logic/aggregationbuilder/AggregationWidgetConfig';
+import Series from 'views/logic/aggregationbuilder/Series';
 
 import AttributeConfigurationContainer from './AttributeConfigurationContainer';
 
-import type { CreateAggregationAttribute } from '../AggregationWizard';
+import type { CreateAggregationAction } from '../AggregationWizard';
 
-export const createAggregationAttribute: CreateAggregationAttribute = (config, onConfigChange) => ({
-  label: 'Sort',
-  value: 'sort',
-  isAvailable: isEmpty(config.sort),
-  onCreate: (config) => {},
-  onDeleteAll: (config) => {},
+export const createAggregationAction: CreateAggregationAction = (config, onConfigChange) => ({
+  label: 'Metric',
+  value: 'metric',
+  isAvailable: isEmpty(config.series),
+  onCreate: () => onConfigChange(config.toBuilder().series([Series.createDefault()]).build()),
+  onDeleteAll: () => onConfigChange(config.toBuilder().series([]).build()),
 });
 
 type Props = {
@@ -37,14 +38,15 @@ type Props = {
   onConfigChange: (newConfig: AggregationWidgetConfig) => void
 }
 
-const SortConfiguration = ({ config, onConfigChange }: Props) => {
-  const aggregationAttribute = createAggregationAttribute(config, onConfigChange);
+const MetricConfiguration = ({ config, onConfigChange }: Props) => {
+  console.log(config);
+  const aggregationAction = createAggregationAction(config, onConfigChange);
 
   return (
-    <AttributeConfigurationContainer aggregationAttribute={aggregationAttribute}>
+    <AttributeConfigurationContainer aggregationAction={aggregationAction}>
       Configuration Elements
     </AttributeConfigurationContainer>
   );
 };
 
-export default SortConfiguration;
+export default MetricConfiguration;
